Fix: Change relevance to double on peers info dto

+++ lib/src/v2/sync/dto/peers_info.dart
@@ -66,7 +66,7 @@ class PeersInfo {
   final double? progress;
 
   @JsonKey(name: 'relevance')
-  final int? relevance;
+  final double? relevance;
 
   @JsonKey(name: 'up_speed')
   final int? upSpeed;
